PIM-4846: move Akeneo\Component\FileStorage\RawFile namespace to Akeneo\Component\FileStorage\File
