+++ src/Akeneo/Bundle/FileStorageBundle/Command/StoreFileCommand.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Bundle\FileStorageBundle\Command;
 
-use Akeneo\Component\FileStorage\RawFile\RawFileStorerInterface;
+use Akeneo\Component\FileStorage\File\RawFileStorerInterface;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
