+++ src/Akeneo/Component/FileStorage/File/RawFileFetcherInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Component\FileStorage\RawFile;
+namespace Akeneo\Component\FileStorage\File;
 
 use Akeneo\Component\FileStorage\Exception\FileTransferException;
 use League\Flysystem\FilesystemInterface;
