+++ src/Pim/Bundle/TransformBundle/Denormalizer/Flat/ProductValue/FileDenormalizer.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\TransformBundle\Denormalizer\Flat\ProductValue;
 
-use Akeneo\Component\FileStorage\RawFile\RawFileStorerInterface;
+use Akeneo\Component\FileStorage\File\RawFileStorerInterface;
 use Akeneo\Component\FileStorage\Repository\FileRepositoryInterface;
 use Pim\Component\Catalog\FileStorage;
