+++ src/Pim/Bundle/TransformBundle/Transformer/Property/MediaTransformer.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\TransformBundle\Transformer\Property;
 
 use Akeneo\Component\FileStorage\Exception\FileTransferException;
-use Akeneo\Component\FileStorage\RawFile\RawFileStorerInterface;
+use Akeneo\Component\FileStorage\File\RawFileStorerInterface;
 use Pim\Bundle\TransformBundle\Exception\PropertyTransformerException;
 use Pim\Bundle\TransformBundle\Transformer\ColumnInfo\ColumnInfoInterface;
 use Pim\Component\Catalog\FileStorage;
